TP-63 게시물 필터링 API
repo, service 인터페이스 분리
